Move IHttpClient to the common types.
