+++ src/client/activation/languageServer/downloader.ts
@@ -10,14 +10,14 @@ import { IApplicationShell, IWorkspaceService } from '../../common/application/t
 import { STANDARD_OUTPUT_CHANNEL } from '../../common/constants';
 import '../../common/extensions';
 import { IFileSystem } from '../../common/platform/types';
-import { IOutputChannel, Resource } from '../../common/types';
+import { IHttpClient, IOutputChannel, Resource } from '../../common/types';
 import { createDeferred } from '../../common/utils/async';
 import { Common, LanguageService } from '../../common/utils/localize';
 import { StopWatch } from '../../common/utils/stopWatch';
 import { sendTelemetryEvent } from '../../telemetry';
 import { EventName } from '../../telemetry/constants';
 import {
-    IHttpClient, ILanguageServerDownloader, ILanguageServerFolderService,
+    ILanguageServerDownloader, ILanguageServerFolderService,
     IPlatformData
 } from '../types';
