+++ src/client/common/net/httpClient.ts
@@ -5,7 +5,7 @@
 
 import { inject, injectable } from 'inversify';
 import * as requestTypes from 'request';
-import { IHttpClient } from '../../activation/types';
+import { IHttpClient } from '../../common/types';
 import { IServiceContainer } from '../../ioc/types';
 import { IWorkspaceService } from '../application/types';
